# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signup successful",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "summary": "Sign up a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "summary": "Login user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "summary": "Refresh JWT token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "summary": "Verify email",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset link sent",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "summary": "Request password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/setup": {
      "post": {
        "operationId": "AuthController_setup2FA",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFASetupDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA QR code and secret",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Setup 2FA (returns QR code)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/enable": {
      "post": {
        "operationId": "AuthController_enable2FA",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFAVerifyDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA enabled",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable 2FA (verify code)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/disable": {
      "post": {
        "operationId": "AuthController_disable2FA",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFAVerifyDto"
              },
              "examples": {
                "user": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA disabled",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable 2FA (verify code)",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS SaaS API",
    "description": "API documentation for the LMS SaaS platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password (min 6 chars)",
            "example": "password123"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          },
          "code": {
            "type": "string",
            "description": "2FA code (if enabled)",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "TwoFASetupDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "User password for 2FA setup",
            "example": "password123"
          }
        },
        "required": [
          "password"
        ]
      },
      "TwoFAVerifyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "2FA verification code",
            "example": "123456"
          }
        },
        "required": [
          "code"
        ]
      },
...
    }
  }
}